Fix bug in task cancellation
The following sequence was reproduced:
Task #1 starts
Task #2 tries to start and issues cancellation of task #1
Task #3 tries to start and issues cancellation of task #1
...

It is wrong by design.
Solution: any task which is going to start, issues cancellation of all previous tasks.
When a task is cancelled, it throws cancellation exception which is handled by application.